+++ lib/interfaces/DownloadTranslationParams.ts
@@ -1,7 +1,9 @@
 import type { DownloadFileParams } from "@lokalise/node-api";
 import type { ExtractParams } from "./ExtractParams.js";
+import type { ProcessDownloadFileParams } from "./ProcessDownloadFileParams.js";
 
 export interface DownloadTranslationParams {
 	downloadFileParams: DownloadFileParams;
 	extractParams?: ExtractParams;
+	processDownloadFileParams?: ProcessDownloadFileParams;
 }