+++ lib/interfaces/ProcessDownloadFileParams.ts
@@ -1,0 +1,5 @@
+export interface ProcessDownloadFileParams {
+	asyncDownload?: boolean;
+	pollInitialWaitTime?: number;
+	pollMaximumWaitTime?: number;
+}